chore: refactor, split resolve-deps.go to pkgs
